organize UI fragments into Components
